Added client-side validation for RateLimit parameters

+++ Web.Portal/Static/libs/weebreader/admin/chapter-editor-comic.js
@@ -24,7 +24,7 @@
                         prompt: chapterMustHaveNumberLabel
                     },
                     {
-                        type: "integer",
+                        type: "decimal",
                         prompt: chapterNumberOutOfRangeLabel
                     }
                 ]
