+++ Web.Portal/Static/libs/weebreader/admin/chapter-editor-novel.js
@@ -22,7 +22,7 @@
                         prompt: chapterMustHaveNumberLabel
                     },
                     {
-                        type: "integer",
+                        type: "decimal",
                         prompt: chapterNumberOutOfRangeLabel
                     }
                 ]
